+++ lib/src/validators/validators.dart
@@ -1,3 +1,4 @@
 export 'banker.dart';
+export 'general.dart';
 export 'national_id.dart';
 export 'phone.dart';